requirements.txt
flask
mysql-connector-python
sqids
python-dotenv